tuấn upatde sửa route admin và user, gắn link, sửa 1 vài bug, xem mật…

+++ DATN_Client/Areas/Admin/Components/Categories.razor.cs
@@ -1,4 +1,5 @@
-﻿using DATN_Shared.ViewModel;
+﻿using DATN_Client.Areas.Customer.Component;
+using DATN_Shared.ViewModel;
 using Microsoft.AspNetCore.Components;
 
 namespace DATN_Client.Areas.Admin.Components
@@ -8,33 +9,75 @@ namespace DATN_Client.Areas.Admin.Components
         HttpClient _httpClient = new HttpClient();
         public Categories_VM cate_VM = new Categories_VM();
         [Inject] NavigationManager navigationManager { get; set; }
-        List<Categories_VM> cate = new List<Categories_VM>();
+        [Inject] private Blazored.Toast.Services.IToastService _toastService { get; set; } // Khai báo khi cần gọi ở code-behind
+        List<Categories_VM> _lstcate = new List<Categories_VM>();
         public string Message { get; set; } = string.Empty;
         protected override async Task OnInitializedAsync()
         {
-            cate = await _httpClient.GetFromJsonAsync<List<Categories_VM>>("https://localhost:7141/api/Categories");
+            if (Login.Roleuser != "Admin")
+            {
+                navigationManager.NavigateTo("https://localhost:7075/Admin", true);
+                return;
+            }
+            _lstcate = await _httpClient.GetFromJsonAsync<List<Categories_VM>>("https://localhost:7141/api/Categories");
         }
         public async Task AddCate()
         {
+            if (Login.Roleuser != "Admin")
+            {
+                navigationManager.NavigateTo("https://localhost:7075/Admin", true);
+                return;
+            }
             cate_VM.Id = Guid.NewGuid();
-
+            if (!string.IsNullOrEmpty(cate_VM.Name))
+            {
+                if (_lstcate.Any(c => c.Name == cate_VM.Name))
+                {
+                    _toastService.ShowError("Thể loại đã tồn tại");
+                    return;
+                }
+            }
+            if (string.IsNullOrEmpty(cate_VM.Name)) return;
             await _httpClient.PostAsJsonAsync<Categories_VM>("https://localhost:7141/api/Categories/PostCategory", cate_VM);
-            navigationManager.NavigateTo("https://localhost:7075/Admin/Categories", true);
-
-
+            navigationManager.NavigateTo("/category-management", true);
         }
         public async Task UpdateCate(Categories_VM cate)
         {
+            if (Login.Roleuser != "Admin")
+            {
+                navigationManager.NavigateTo("https://localhost:7075/Admin", true);
+                return;
+            }
+            if (!string.IsNullOrEmpty(cate_VM.Name))
+            {
+                if (_lstcate.Any(c => c.Name == cate_VM.Name))
+                {
+                    _toastService.ShowError("Thể loại đã tồn tại");
+                    return;
+                }
+            }
+            if (string.IsNullOrEmpty(cate_VM.Name)) return;
+            cate.TenKhongDau = "";
             await _httpClient.PutAsJsonAsync<Categories_VM>("https://localhost:7141/api/Categories/PutCategory", cate);
-            navigationManager.NavigateTo("https://localhost:7075/Admin/Categories", true);
+            navigationManager.NavigateTo("/category-management", true);
         }
         public async void DeleteCate(Guid Id)
         {
+            if (Login.Roleuser != "Admin")
+            {
+                navigationManager.NavigateTo("https://localhost:7075/Admin", true);
+                return;
+            }
             await _httpClient.DeleteAsync("https://localhost:7141/api/Categories/DeleteCategory/" + Id);
-            navigationManager.NavigateTo("https://localhost:7075/Admin/Categories", true);
+            navigationManager.NavigateTo("/category-management", true);
         }
         public async Task LoadForm(Categories_VM rvm)
         {
+            if (Login.Roleuser != "Admin")
+            {
+                navigationManager.NavigateTo("https://localhost:7075/Admin", true);
+                return;
+            }
             cate_VM.Id = rvm.Id;
             cate_VM.Name = rvm.Name;
             cate_VM.Status = rvm.Status;
